data for tennis
here 8 input data, in which one 4 tests and 4 mst clasters for k = 4, if
u see that k = 4 is bad, tell me, i create new mst clasters for ur k. u
need to calculate error for test1 using source1 from this folder. in
source the first number is index, the second is number of claster